+++ core/src/main/java/kuburger/domain/model/PaymentStatus.java
@@ -1,0 +1,19 @@
+package kuburger.domain.model;
+
+public enum PaymentStatus {
+    PENDING("Pending"), PAID("Paid");
+
+    private String value;
+
+    PaymentStatus(String value) {
+        this.value = value;
+    }
+
+    public String getCode() {
+        return super.toString();
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
