+++ core/src/main/java/kuburger/domain/model/Product.java
@@ -1,0 +1,48 @@
+package kuburger.domain.model;
+
+import java.io.Serializable;
+import java.util.Objects;
+
+public abstract class Product implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    private Long id;
+    private Double price;
+    private String description;
+
+    public void setPrice(Double price) {
+        this.price = price;
+    }
+
+    public Double getPrice() {
+        return price;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Product product = (Product) o;
+        return Double.compare(product.price, price) == 0 &&
+                Objects.equals(id, product.id) &&
+                Objects.equals(description, product.description);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, price, description);
+    }
+}
