--- src/main/java/com/delivery/kuburger/KuburgerApplication.java
@@ -1,14 +1,0 @@
-package com.delivery.kuburger;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KuburgerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(KuburgerApplication.class, args);
-    }
-
-}
-
