--- src/test/java/com/delivery/kuburger/KuburgerApplicationTests.java
@@ -1,17 +1,0 @@
-package com.delivery.kuburger;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class KuburgerApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
-
