--- src/test/java/com/delivery/kuburger/entity/AddressTest.java
@@ -1,22 +1,0 @@
-package com.delivery.kuburger.entity;
-
-import junitx.extensions.EqualsHashCodeTestCase;
-
-public class AddressTest extends EqualsHashCodeTestCase {
-
-    public AddressTest(String name) {
-        super(name);
-    }
-
-    @Override
-    protected Object createInstance() throws Exception {
-        return new Address();
-    }
-
-    @Override
-    protected Object createNotEqualInstance() throws Exception {
-        Address address = new Address();
-        address.setCity("Fortaleza");
-        return address;
-    }
-}
